Refactor PDFReconApp methods to scanner module
