+++ pdfrecon/app_gui.py
@@ -58,6 +58,19 @@ requests = _import_with_fallback('requests', 'requests', 'requests')
 from .config import PDFReconConfig, PDFProcessingError, PDFCorruptionError, \
     PDFTooLargeError, PDFEncryptedError, APP_VERSION, UI_COLORS, UI_FONTS, UI_DIMENSIONS
 
+from .scanner import (
+    process_single_file,
+    scan_worker_parallel,
+    generate_comprehensive_timeline,
+    run_exiftool,
+    extract_text_for_indicators,
+    detect_indicators,
+    extract_revisions,
+    extract_additional_xmp_ids,
+    get_text_for_comparison
+)
+from .utils import resolve_path
+
 # --- OCG (layers) detection helpers ---
 _LAYER_OCGS_BLOCK_RE = re.compile(rb"/OCGs\s*\[(.*?)\]", re.S)
 _OBJ_REF_RE          = re.compile(rb"(\d+)\s+(\d+)\s+R")
@@ -158,7 +171,7 @@ class PDFReconApp:
         self.is_reader_mode = "reader" in Path(sys.executable).stem.lower()
         
         # --- Configuration Setup ---
-        self.config_path = self._resolve_path("config.ini", base_is_parent=True)
+        self.config_path = resolve_path("config.ini", base_is_parent=True)
         self._load_or_create_config()
         
         # --- Window Configuration ---
@@ -236,7 +249,7 @@ class PDFReconApp:
         
         # --- Set Application Icon ---
         try:
-            icon_path = self._resolve_path('icon.ico')
+            icon_path = resolve_path('icon.ico')
             if icon_path.exists():
                 self.root.iconbitmap(icon_path)
             else:
@@ -533,7 +546,7 @@ class PDFReconApp:
             
         # Try multiple locations for the log file
         log_locations = [
-            self._resolve_path("pdfrecon.log", base_is_parent=True),  # App directory
+            resolve_path("pdfrecon.log", base_is_parent=True),  # App directory
             Path(tempfile.gettempdir()) / "pdfrecon.log",  # Temp directory
         ]
         
@@ -1622,32 +1635,6 @@ class PDFReconApp:
         text_widget.bind("<Control-c>", copy_selection) # Ctrl+C
         text_widget.bind("<Command-c>", copy_selection) # Command+C for macOS
         
-    def _add_layer_indicators(self, raw: bytes, path: Path, indicators: dict):
-        """
-        Adds indicators for layers:
-          - "Has Layers (count)" if OCGs are found.
-          - "More Layers Than Pages" if layer count > page count.
-        """
-        try:
-            layers_cnt = count_layers(raw)
-        except Exception:
-            layers_cnt = 0
-
-        if layers_cnt <= 0:
-            return
-
-        indicators['HasLayers'] = {'count': layers_cnt}
-
-        # Compare with page count (best-effort)
-        page_count = 0
-        try:
-            with fitz.open(path) as _doc:
-                page_count = _doc.page_count
-        except Exception:
-            pass
-
-        if page_count and layers_cnt > page_count:
-            indicators['MoreLayersThanPages'] = {'layers': layers_cnt, 'pages': page_count}
     def _populate_timeline_widget(self, text_widget, path_str):
         """Helper function to populate a text widget with formatted timeline data."""
         timeline_data = self.timeline_data.get(path_str)
@@ -2173,29 +2160,6 @@ class PDFReconApp:
 
 
         
-    def _resolve_path(self, filename, base_is_parent=False):
-        """
-        Resolves the correct path for a resource file, whether running as a script
-        or as a frozen executable (e.g., with PyInstaller).
-        """
-        if getattr(sys, 'frozen', False):
-            # If the app is frozen, the base path is the folder containing the exe.
-            base_path = Path(sys.executable).parent
-            # When frozen and base_is_parent=True, look for the file next to the exe (e.g., exiftool.exe)
-            if base_is_parent:
-                return base_path / filename
-            # Otherwise, look in _MEIPASS for bundled data files (translations, etc.)
-            return Path(getattr(sys, '_MEIPASS', base_path)) / filename
-        else:
-            # If running as a normal script, the base path is the script's folder.
-            base_path = Path(__file__).resolve().parent
-        
-        # If base_is_parent=True, go up one level to project root
-        if base_is_parent:
-            base_path = base_path.parent
-        
-        return base_path / filename
-
     def show_license(self):
         """Displays the license information from 'license.txt' in a popup."""
         # Try multiple locations for license.txt
@@ -2306,159 +2270,14 @@ class PDFReconApp:
         self.copy_executor = ThreadPoolExecutor(max_workers=4, thread_name_prefix='CopyWorker')
 
         # --- Start Main Worker Thread ---
-        scan_thread = threading.Thread(target=self._scan_worker_parallel, args=(folder_path, self.scan_queue))
+        scan_thread = threading.Thread(
+            target=scan_worker_parallel,
+            args=(folder_path, self.scan_queue, self.copy_executor, self._)
+        )
         scan_thread.daemon = True
         scan_thread.start()
 
         self._process_queue()
-
-    def _find_pdf_files_generator(self, folder):
-        """Generator that yields PDF files as soon as they are found."""
-        for base, _, files in os.walk(folder):
-            for fn in files:
-                if fn.lower().endswith(".pdf"):
-                    yield Path(base) / fn
-
-    def _process_single_file(self, fp):
-        """
-        Processes a single PDF file - analyzes it and extracts metadata.
-        Returns a list of dictionaries (one for original, one per revision).
-        """
-        try:
-            # Validate the file first
-            file_size = fp.stat().st_size
-            if file_size > PDFReconConfig.MAX_FILE_SIZE:
-                raise PDFTooLargeError(f"File size {file_size / (1024**2):.1f}MB exceeds limit")
-            
-            # Read file and open PDF
-            raw = fp.read_bytes()
-            doc = self._safe_pdf_open(fp, raw_bytes=raw)
-            
-            # Extract text for indicator detection
-            txt = self.extract_text(raw)
-            
-            # === CRITICAL: Run ExifTool HERE ===
-            exif = self.exiftool_output(fp, detailed=True)
-            
-            # Extract document IDs for cross-referencing
-            document_ids = self._extract_all_document_ids(txt, exif)
-            
-            # Detect indicators
-            indicator_keys = self.detect_indicators(fp, txt, doc)
-            
-            # Calculate MD5
-            md5_hash = hashlib.md5(raw).hexdigest()
-            
-            # Generate timeline
-            original_timeline = self.generate_comprehensive_timeline(fp, txt, exif)
-            
-            # Extract revisions
-            revisions = self.extract_revisions(raw, fp)
-            
-            # Close the document
-            doc.close()
-            
-            # Add "HasRevisions" indicator if revisions were found
-            final_indicator_keys = indicator_keys.copy()
-            if revisions:
-                final_indicator_keys['HasRevisions'] = {'count': len(revisions)}
-            
-            # Prepare result for original file
-            results = []
-            original_row_data = {
-                "path": fp,
-                "indicator_keys": final_indicator_keys,
-                "md5": md5_hash,
-                "exif": exif,
-                "is_revision": False,
-                "timeline": original_timeline,
-                "status": "success",
-                "document_ids": document_ids
-            }
-            results.append(original_row_data)
-            
-            # Process each revision
-            for rev_path, basefile, rev_raw in revisions:
-                try:
-                    rev_md5 = hashlib.md5(rev_raw).hexdigest()
-                    rev_exif = self.exiftool_output(rev_path, detailed=True)
-                    rev_txt = self.extract_text(rev_raw)
-                    revision_timeline = self.generate_comprehensive_timeline(rev_path, rev_txt, rev_exif)
-                    
-                    revision_row_data = {
-                        "path": rev_path,
-                        "indicator_keys": {"Revision": {}},
-                        "md5": rev_md5,
-                        "exif": rev_exif,
-                        "is_revision": True,
-                        "timeline": revision_timeline,
-                        "original_path": fp,
-                        "is_identical": False,
-                        "status": "success"
-                    }
-                    results.append(revision_row_data)
-                except Exception as e:
-                    logging.warning(f"Error processing revision {rev_path.name}: {e}")
-            
-            return results
-            
-        except PDFTooLargeError as e:
-            logging.warning(f"Skipping large file {fp.name}: {e}")
-            return [{"path": fp, "status": "error", "error_type": "file_too_large", "error_message": str(e)}]
-        except PDFEncryptedError as e:
-            logging.warning(f"Skipping encrypted file {fp.name}: {e}")
-            return [{"path": fp, "status": "error", "error_type": "file_encrypted", "error_message": str(e)}]
-        except PDFCorruptionError as e:
-            logging.warning(f"Skipping corrupt file {fp.name}: {e}")
-            return [{"path": fp, "status": "error", "error_type": "file_corrupt", "error_message": str(e)}]
-        except Exception as e:
-            logging.exception(f"Unexpected error processing file {fp.name}")
-            return [{"path": fp, "status": "error", "error_type": "processing_error", "error_message": str(e)}]
-
-    def _scan_worker_parallel(self, folder, q):
-        """
-        Worker thread that finds and processes PDF files in parallel.
-        Sends results to the queue for UI updates.
-        """
-        try:
-            q.put(("scan_status", self._("preparing_analysis")))
-            
-            # Find all PDF files
-            pdf_files = list(self._find_pdf_files_generator(folder))
-            if not pdf_files:
-                q.put(("finished", None))
-                return
-            
-            q.put(("progress_mode_determinate", len(pdf_files)))
-            files_processed = 0
-            
-            # Process files in parallel using ThreadPoolExecutor
-            with ThreadPoolExecutor(max_workers=PDFReconConfig.MAX_WORKER_THREADS) as executor:
-                future_to_path = {executor.submit(self._process_single_file, fp): fp for fp in pdf_files}
-                
-                for future in as_completed(future_to_path):
-                    path = future_to_path[future]
-                    files_processed += 1
-                    
-                    try:
-                        results = future.result()
-                        for result_data in results:
-                            q.put(("file_row", result_data))
-                    except Exception as e:
-                        logging.error(f"Unexpected error from thread pool for file {path.name}: {e}")
-                        q.put(("file_row", {"path": path, "status": "error", "error_type": "unknown_error", "error_message": str(e)}))
-                    
-                    # Calculate progress stats
-                    elapsed_time = time.time() - self.scan_start_time
-                    fps = files_processed / elapsed_time if elapsed_time > 0 else 0
-                    eta_seconds = (len(pdf_files) - files_processed) / fps if fps > 0 else 0
-                    q.put(("detailed_progress", {"file": path.name, "fps": fps, "eta": time.strftime('%M:%S', time.gmtime(eta_seconds))}))
-        
-        except Exception as e:
-            logging.error(f"Error in scan worker: {e}")
-            q.put(("error", f"A critical error occurred: {e}"))
-        finally:
-            q.put(("finished", None))
 
     def _process_queue(self):
         """
@@ -2877,7 +2696,7 @@ class PDFReconApp:
 
             dependencies = ["license.txt", "config.ini", "icon.ico"]
             for dep_name in dependencies:
-                source_dep = self._resolve_path(dep_name, base_is_parent=True)
+                source_dep = resolve_path(dep_name, base_is_parent=True)
                 if source_dep.exists():
                     failed_file = dep_name
                     shutil.copy2(source_dep, dest_folder / dep_name)
@@ -2893,253 +2712,6 @@ class PDFReconApp:
                 self._("export_reader_error_specific_msg").format(filename=failed_file, e=e)
             )
             
-    def _find_pdf_files_generator(self, folder):
-        """A generator that 'yields' PDF files as soon as they are found."""
-        for base, _, files in os.walk(folder):
-            for fn in files:
-                if fn.lower().endswith(".pdf"):
-                    yield Path(base) / fn
-
-    def validate_pdf_file(self, filepath):
-        """
-        Validates a PDF file based on size, header, and encryption.
-        Returns True if valid, otherwise raises an appropriate exception.
-        """
-        try:
-            # Check file size
-            if filepath.stat().st_size > PDFReconConfig.MAX_FILE_SIZE:
-                raise PDFTooLargeError(f"File exceeds {PDFReconConfig.MAX_FILE_SIZE // (1024*1024)}MB size limit.")
-
-            # Check PDF header
-            with open(filepath, 'rb') as f:
-                if f.read(5) != b'%PDF-':
-                    raise PDFCorruptionError("Invalid PDF header. Not a PDF file.")
-            
-            # Check for encryption
-            with fitz.open(filepath) as doc:
-                if doc.is_encrypted:
-                    # Try authenticating with an empty password
-                    if not doc.authenticate(""):
-                         raise PDFEncryptedError("File is encrypted and cannot be processed.")
-            
-            return True
-        except PDFProcessingError:
-            # Re-raise our custom exceptions
-            raise
-        except Exception as e:
-            # Wrap other exceptions in our custom error type
-            raise PDFCorruptionError(f"Could not validate file: {e}")
-
-    def _process_large_file_streaming(self, filepath):
-        """(Placeholder) Processes a large PDF file using streaming to save memory."""
-        logging.info(f"Streaming processing is not yet implemented. Processing {filepath.name} normally.")
-        pass
-    def extract_additional_xmp_ids(self, txt: str) -> dict:
-        """
-        Harvest XMP IDs beyond basic xmpMM:{Original,Document,Instance}:
-        - stRef:documentID / stRef:instanceID   (both attribute AND element forms)
-        - xmpMM:DerivedFrom   (stRef:* and OriginalDocumentID inside)
-        - xmpMM:Ingredients   (stRef:* inside)
-        - xmpMM:History       (stRef:* and any InstanceID-like values)
-        - photoshop:DocumentAncestors (Bag of GUIDs/uuids)
-        Returns a dict of sets with normalized uppercase IDs.
-        """
-        def _norm(val):
-            """Normalizes a UUID/GUID value for consistent comparison."""
-            if val is None:
-                return None
-            if isinstance(val, (bytes, bytearray)):
-                val = val.decode("utf-8", "ignore")
-            s = str(val).strip()
-            # strip known prefixes & wrappers
-            s = re.sub(r"^urn:uuid:", "", s, flags=re.I)
-            s = re.sub(r"^(uuid:|xmp\.iid:|xmp\.did:)", "", s, flags=re.I)
-            s = s.strip("<>").strip()
-            return s.upper() if s else None
-
-        out = {
-            "stref_doc_ids": set(),
-            "stref_inst_ids": set(),
-            "derived_doc_ids": set(),
-            "derived_inst_ids": set(),
-            "derived_orig_ids": set(),
-            "ingredient_doc_ids": set(),
-            "ingredient_inst_ids": set(),
-            "history_inst_ids": set(),
-            "history_doc_ids": set(),
-            "ps_doc_ancestors": set(),
-        }
-
-        # ---------------- stRef anywhere (attribute form) ----------------
-        for m in re.finditer(r'stRef:documentID="([^"]+)"', txt, re.I):
-            v = _norm(m.group(1));  out["stref_doc_ids"].add(v) if v else None
-        for m in re.finditer(r'stRef:instanceID="([^"]+)"', txt, re.I):
-            v = _norm(m.group(1));  out["stref_inst_ids"].add(v) if v else None
-
-        # ---------------- stRef anywhere (element form) ----------------
-        for m in re.finditer(r"<stRef:documentID>([^<]+)</stRef:documentID>", txt, re.I):
-            v = _norm(m.group(1));  out["history_doc_ids"].add(v) if v else None
-        for m in re.finditer(r"<stRef:instanceID>([^<]+)</stRef:instanceID>", txt, re.I):
-            v = _norm(m.group(1));  out["history_inst_ids"].add(v) if v else None
-
-        # ---------------- DerivedFrom block ----------------
-        df = re.search(r"<xmpMM:DerivedFrom\b[^>]*>(.*?)</xmpMM:DerivedFrom>", txt, re.I | re.S)
-        if df:
-            blk = df.group(1)
-            # stRef:* within DerivedFrom
-            for m in re.finditer(r'stRef:documentID="([^"]+)"', blk, re.I):
-                v = _norm(m.group(1)); out["derived_doc_ids"].add(v) if v else None
-            for m in re.finditer(r'stRef:instanceID="([^"]+)"', blk, re.I):
-                v = _norm(m.group(1)); out["derived_inst_ids"].add(v) if v else None
-            for m in re.finditer(r"<stRef:documentID>([^<]+)</stRef:documentID>", blk, re.I):
-                v = _norm(m.group(1)); out["derived_doc_ids"].add(v) if v else None
-            for m in re.finditer(r"<stRef:instanceID>([^<]+)</stRef:instanceID>", blk, re.I):
-                v = _norm(m.group(1)); out["derived_inst_ids"].add(v) if v else None
-            # OriginalDocumentID sometimes appears explicitly inside DerivedFrom
-            for m in re.finditer(r"(?:xmpMM:|)OriginalDocumentID(?:>|=\")([^<\">]+)", blk, re.I):
-                v = _norm(m.group(1)); out["derived_orig_ids"].add(v) if v else None
-
-        # ---------------- Ingredients block ----------------
-        ing = re.search(r"<xmpMM:Ingredients\b[^>]*>(.*?)</xmpMM:Ingredients>", txt, re.I | re.S)
-        if ing:
-            blk = ing.group(1)
-            for m in re.finditer(r'stRef:documentID="([^"]+)"', blk, re.I):
-                v = _norm(m.group(1)); out["ingredient_doc_ids"].add(v) if v else None
-            for m in re.finditer(r'stRef:instanceID="([^"]+)"', blk, re.I):
-                v = _norm(m.group(1)); out["ingredient_inst_ids"].add(v) if v else None
-            for m in re.finditer(r"<stRef:documentID>([^<]+)</stRef:documentID>", blk, re.I):
-                v = _norm(m.group(1)); out["ingredient_doc_ids"].add(v) if v else None
-            for m in re.finditer(r"<stRef:instanceID>([^<]+)</stRef:instanceID>", blk, re.I):
-                v = _norm(m.group(1)); out["ingredient_inst_ids"].add(v) if v else None
-
-        # ---------------- History block ----------------
-        hist = re.search(r"<xmpMM:History\b[^>]*>(.*?)</xmpMM:History>", txt, re.I | re.S)
-        if hist:
-            blk = hist.group(1)
-            # attribute-style summary sometimes exists in certain producers
-            for m in re.finditer(r'(?:InstanceID|stRef:instanceID)="([^"]+)"', blk, re.I):
-                v = _norm(m.group(1)); out["history_inst_ids"].add(v) if v else None
-            for m in re.finditer(r'(?:DocumentID|stRef:documentID)="([^"]+)"', blk, re.I):
-                v = _norm(m.group(1)); out["history_doc_ids"].add(v) if v else None
-            # element-style
-            for m in re.finditer(r"<stRef:instanceID>([^<]+)</stRef:instanceID>", blk, re.I):
-                v = _norm(m.group(1)); out["history_inst_ids"].add(v) if v else None
-            for m in re.finditer(r"<stRef:documentID>([^<]+)</stRef:documentID>", blk, re.I):
-                v = _norm(m.group(1)); out["history_doc_ids"].add(v) if v else None
-            # catch any xmp.iid/xmp.did that appear as plain text within History
-            for m in re.finditer(r"(uuid:[0-9a-f\-]+|xmp\.iid:[^,<>} \]]+|xmp\.did:[^,<>} \]]+)", blk, re.I):
-                v = _norm(m.group(1)); out["history_inst_ids"].add(v) if v else None
-
-        # ---------------- Photoshop DocumentAncestors ----------------
-        ps = re.search(r"<photoshop:DocumentAncestors\b[^>]*>(.*?)</photoshop:DocumentAncestors>", txt, re.I | re.S)
-        if ps:
-            for m in re.finditer(r"<rdf:li[^>]*>([^<]+)</rdf:li>", ps.group(1), re.I):
-                v = _norm(m.group(1)); out["ps_doc_ancestors"].add(v) if v else None
-
-        # Deduplicate and purge empty values
-        for k in out:
-            out[k] = {v for v in out[k] if v}
-
-        return out
-
-    def _extract_all_document_ids(self, txt: str, exif_output: str) -> dict:
-        """
-        Extracts all document IDs from a PDF for cross-referencing.
-        Returns a dict with 'own_ids' (this file's identifiers) and 'ref_ids' (references to other documents).
-        """
-        def _norm(val):
-            if val is None:
-                return None
-            if isinstance(val, (bytes, bytearray)):
-                val = val.decode("utf-8", "ignore")
-            s = str(val).strip()
-            s = re.sub(r"^urn:uuid:", "", s, flags=re.I)
-            s = re.sub(r"^(uuid:|xmp\.iid:|xmp\.did:)", "", s, flags=re.I)
-            s = s.strip("<>").strip()
-            return s.upper() if s else None
-
-        own_ids = set()  # IDs that identify THIS document
-        ref_ids = set()  # IDs that reference OTHER documents (ancestors, derived from, etc.)
-
-        # Extract this document's own IDs
-        # XMP DocumentID
-        m = re.search(r'xmpMM:DocumentID(?:>|=")([^<"]+)', txt, re.I)
-        if m:
-            v = _norm(m.group(1))
-            if v: own_ids.add(v)
-
-        # XMP InstanceID
-        m = re.search(r'xmpMM:InstanceID(?:>|=")([^<"]+)', txt, re.I)
-        if m:
-            v = _norm(m.group(1))
-            if v: own_ids.add(v)
-
-        # PDF Trailer IDs (first and second)
-        for m in re.finditer(r"/ID\s*\[\s*<([0-9A-Fa-f]+)>\s*<([0-9A-Fa-f]+)>\s*\]", txt):
-            v1, v2 = _norm(m.group(1)), _norm(m.group(2))
-            if v1: own_ids.add(v1)
-            if v2: own_ids.add(v2)
-
-        # Extract referenced IDs (ancestors, derived from, etc.)
-        # XMP OriginalDocumentID (reference to original)
-        m = re.search(r'xmpMM:OriginalDocumentID(?:>|=")([^<"]+)', txt, re.I)
-        if m:
-            v = _norm(m.group(1))
-            if v: ref_ids.add(v)
-
-        # DerivedFrom block
-        df = re.search(r"<xmpMM:DerivedFrom\b[^>]*>(.*?)</xmpMM:DerivedFrom>", txt, re.I | re.S)
-        if df:
-            blk = df.group(1)
-            for m in re.finditer(r'stRef:documentID(?:>|=")([^<"]+)', blk, re.I):
-                v = _norm(m.group(1))
-                if v: ref_ids.add(v)
-            for m in re.finditer(r'stRef:instanceID(?:>|=")([^<"]+)', blk, re.I):
-                v = _norm(m.group(1))
-                if v: ref_ids.add(v)
-
-        # Ingredients block (embedded/linked documents)
-        ing = re.search(r"<xmpMM:Ingredients\b[^>]*>(.*?)</xmpMM:Ingredients>", txt, re.I | re.S)
-        if ing:
-            blk = ing.group(1)
-            for m in re.finditer(r'stRef:documentID(?:>|=")([^<"]+)', blk, re.I):
-                v = _norm(m.group(1))
-                if v: ref_ids.add(v)
-
-        # Photoshop DocumentAncestors
-        ps = re.search(r"<photoshop:DocumentAncestors\b[^>]*>(.*?)</photoshop:DocumentAncestors>", txt, re.I | re.S)
-        if ps:
-            for m in re.finditer(r"<rdf:li[^>]*>([^<]+)</rdf:li>", ps.group(1), re.I):
-                v = _norm(m.group(1))
-                if v: ref_ids.add(v)
-
-        # History references (past versions)
-        hist = re.search(r"<xmpMM:History\b[^>]*>(.*?)</xmpMM:History>", txt, re.I | re.S)
-        if hist:
-            blk = hist.group(1)
-            for m in re.finditer(r'stRef:documentID(?:>|=")([^<"]+)', blk, re.I):
-                v = _norm(m.group(1))
-                if v: ref_ids.add(v)
-
-        # Also check ExifTool output for any DocumentID/InstanceID
-        if exif_output:
-            for m in re.finditer(r"Document\s*ID\s*:\s*(\S+)", exif_output, re.I):
-                v = _norm(m.group(1))
-                if v: own_ids.add(v)
-            for m in re.finditer(r"Instance\s*ID\s*:\s*(\S+)", exif_output, re.I):
-                v = _norm(m.group(1))
-                if v: own_ids.add(v)
-            for m in re.finditer(r"Original\s*Document\s*ID\s*:\s*(\S+)", exif_output, re.I):
-                v = _norm(m.group(1))
-                if v: ref_ids.add(v)
-
-        # Remove any overlap (don't count own IDs as references)
-        ref_ids = ref_ids - own_ids
-
-        return {
-            "own_ids": own_ids,
-            "ref_ids": ref_ids
-        }
 
     def _cross_reference_document_ids(self):
         """
@@ -3235,64 +2807,6 @@ class PDFReconApp:
         if relationships:
             logging.info(f"Document ID cross-reference found {len(relationships)} files with relationships.")
     
-    def _scan_worker_parallel(self, folder, q):
-        """
-        Finds PDF files and processes them in parallel using a ThreadPoolExecutor.
-        Results are sent back to the main thread via a queue.
-        """
-        try:
-            q.put(("scan_status", self._("preparing_analysis")))
-            
-            pdf_files = list(self._find_pdf_files_generator(folder))
-            if not pdf_files:
-                q.put(("finished", None))
-                return
-
-            q.put(("progress_mode_determinate", len(pdf_files)))
-            files_processed = 0
-
-            # Use a timeout of 60 seconds per file to prevent hangs
-            FILE_TIMEOUT = 60  # seconds
-
-            with ThreadPoolExecutor(max_workers=PDFReconConfig.MAX_WORKER_THREADS) as executor:
-                # Pass the scan root 'folder' to each file processing job
-                future_to_path = {executor.submit(self._process_single_file, fp, folder): fp for fp in pdf_files}
-                
-                # Process completed futures without a global timeout
-                # Use individual future.result() timeouts instead
-                try:
-                    for future in as_completed(future_to_path):
-                        path = future_to_path[future]
-                        files_processed += 1
-                        try:
-                            results = future.result(timeout=FILE_TIMEOUT)
-                            for result_data in results:
-                                q.put(("file_row", result_data))
-                        except TimeoutError:
-                            logging.error(f"TIMEOUT processing file {path.name} - exceeded {FILE_TIMEOUT} seconds")
-                            q.put(("file_row", {"path": path, "status": "error", "error_type": "processing_timeout", "error_message": f"File processing timed out after {FILE_TIMEOUT} seconds"}))
-                        except Exception as e:
-                            logging.error(f"Unexpected error from thread pool for file {path.name}: {e}")
-                            q.put(("file_row", {"path": path, "status": "error", "error_type": "unknown_error", "error_message": str(e)}))
-                        
-                        elapsed_time = time.time() - self.scan_start_time
-                        fps = files_processed / elapsed_time if elapsed_time > 0 else 0
-                        eta_seconds = (len(pdf_files) - files_processed) / fps if fps > 0 else 0
-                        q.put(("detailed_progress", {"file": path.name, "fps": fps, "eta": time.strftime('%M:%S', time.gmtime(eta_seconds))}))
-                except TimeoutError as te:
-                    # Handle case where as_completed itself times out (shouldn't happen now without timeout param)
-                    logging.warning(f"Some futures did not complete: {te}")
-                    # Process any remaining unfinished futures
-                    for future, path in future_to_path.items():
-                        if not future.done():
-                            future.cancel()
-                            q.put(("file_row", {"path": path, "status": "error", "error_type": "cancelled", "error_message": "Processing was cancelled due to timeout"}))
-
-        except Exception as e:
-            logging.error(f"Error in scan worker: {e}")
-            q.put(("error", f"A critical error occurred: {e}"))
-        finally:
-            q.put(("finished", None))
 
     def _process_queue(self):
         """
@@ -3550,271 +3064,6 @@ class PDFReconApp:
                     messagebox.showerror(self._("open_folder_error_title"), f"Could not open the folder: {e}")
                 break
 
-    def extract_revisions(self, raw, original_path):
-        """
-        Extracts previous versions (revisions) of a PDF from its raw byte content
-        by looking for '%%EOF' markers. It prepares potential paths but does not write files.
-        """
-        revisions = []
-        offsets = []
-        pos = len(raw)
-        # Find all '%%EOF' markers from the end of the file backwards
-        while (pos := raw.rfind(b"%%EOF", 0, pos)) != -1: offsets.append(pos)
-        
-        # Filter out invalid or unlikely offsets
-        valid_offsets = [o for o in sorted(offsets) if 1000 <= o <= len(raw) - 500]
-        if valid_offsets:
-            # Define the subdirectory for potential revisions
-            altered_dir = original_path.parent / "Altered_files"
-            
-            for i, offset in enumerate(valid_offsets, start=1):
-                rev_bytes = raw[:offset + 5] # The revision is the content from the start to the EOF marker
-                rev_filename = f"{original_path.stem}_rev{i}_@{offset}.pdf"
-                rev_path = altered_dir / rev_filename
-                # Package the data for later validation without writing the file.
-                revisions.append((rev_path, original_path.name, rev_bytes))
-                
-        return revisions
-
-    def exiftool_output(self, path, detailed=False):
-        """Runs ExifTool safely with a timeout and improved error handling."""
-        exe_path = self._resolve_path("exiftool.exe", base_is_parent=True)
-        if not exe_path.is_file(): return self._("exif_err_notfound")
-        
-        try:
-            file_content = path.read_bytes()
-            # Suppress console window on Windows
-            startupinfo = None
-            if sys.platform == "win32":
-                startupinfo = subprocess.STARTUPINFO()
-                startupinfo.dwFlags |= subprocess.STARTF_USESHOWWINDOW
-            
-            # Build the command-line arguments for ExifTool
-            command = [str(exe_path)]
-            if detailed: command.extend(["-a", "-u", "-s", "-G1", "-struct"])
-            else: command.extend(["-a", "-u", "-s", "-G1"])
-            command.append("-") # Read from stdin
-
-            # Run the process
-            process = subprocess.run(
-                command,
-                input=file_content,
-                capture_output=True,
-                check=False,
-                startupinfo=startupinfo,
-                timeout=PDFReconConfig.EXIFTOOL_TIMEOUT
-            )
-            
-            # Handle non-zero exit codes or stderr output
-            if process.returncode != 0 or process.stderr:
-                error_message = process.stderr.decode('latin-1', 'ignore').strip()
-                if not process.stdout.strip(): return f"{self._('exif_err_prefix')}\n{error_message}"
-                logging.warning(f"ExifTool stderr for {path.name}: {error_message}")
-
-            # Decode the output, trying UTF-8 first, then latin-1 as a fallback
-            try: raw_output = process.stdout.decode('utf-8').strip()
-            except UnicodeDecodeError: raw_output = process.stdout.decode('latin-1', 'ignore').strip()
-
-            # Remove empty lines from the output
-            return "\n".join([line for line in raw_output.splitlines() if line.strip()])
-
-        except subprocess.TimeoutExpired:
-            logging.error(f"ExifTool timed out for file {path.name}")
-            return self._("exif_err_prefix") + f"\nTimeout after {PDFReconConfig.EXIFTOOL_TIMEOUT} seconds."
-        except Exception as e:
-            logging.error(f"Error running exiftool for file {path}: {e}")
-            return self._("exif_err_run").format(e=e)
-
-    def _get_filesystem_times(self, filepath):
-            """Helper function to get created/modified timestamps from the file system."""
-            events = []
-            try:
-                stat = filepath.stat()
-                # Make the datetime object timezone-aware using the system's local timezone
-                mtime = datetime.fromtimestamp(stat.st_mtime).astimezone()
-                events.append((mtime, f"File System: {self._('col_modified')}"))
-                # Make the datetime object timezone-aware using the system's local timezone
-                ctime = datetime.fromtimestamp(stat.st_ctime).astimezone()
-                events.append((ctime, f"File System: {self._('col_created')}"))
-            except FileNotFoundError:
-                pass
-            return events
-            
-    def _parse_exif_data(self, exiftool_output: str):
-        """
-        Parses EXIFTool output into a structured dictionary for reuse.
-        """
-        data = {
-            "producer_pdf": "", "producer_xmppdf": "", "softwareagent": "",
-            "application": "", "software": "", "creatortool": "", "xmptoolkit": "",
-            "create_dt": None, "modify_dt": None, "history_events": [], "all_dates": []
-        }
-        lines = exiftool_output.splitlines()
-
-        # --- Regex Patterns (reuse module-level constants) ---
-        history_pattern = re.compile(r"\[XMP-xmpMM\]\s+History\s+:\s+(.*)")
-
-        def looks_like_software(s: str) -> bool:
-            return bool(s and self.software_tokens.search(s))
-
-        # --- First Pass: Collect Key-Value Pairs for Tools ---
-        for ln in lines:
-            m = _KV_PATTERN.match(ln)
-            if not m: 
-                continue
-            
-            group = m.group("group").strip().lower()
-            tag = m.group("tag").strip().lower().replace(" ", "")
-            val = m.group("value").strip()
-
-            # Map tags to data dictionary keys for software detection
-            if tag == "producer":
-                if group == "pdf" and not data["producer_pdf"]: 
-                    data["producer_pdf"] = val
-                elif group in ("xmp-pdf", "xmp_pdf") and not data["producer_xmppdf"]: 
-                    data["producer_xmppdf"] = val
-            elif tag == "softwareagent" and not data["softwareagent"]: 
-                data["softwareagent"] = val
-            elif tag == "application" and not data["application"]: 
-                data["application"] = val
-            elif tag == "software" and not data["software"]: 
-                data["software"] = val
-            elif tag == "creatortool" and not data["creatortool"] and looks_like_software(val):
-                data["creatortool"] = val
-            elif tag == "xmptoolkit" and not data["xmptoolkit"]: 
-                data["xmptoolkit"] = val
-        
-        # Fallback for producer fields
-        if not data["producer_pdf"] and data["producer_xmppdf"]: 
-            data["producer_pdf"] = data["producer_xmppdf"]
-        if not data["producer_xmppdf"] and data["producer_pdf"]: 
-            data["producer_xmppdf"] = data["producer_pdf"]
-
-        # --- Second Pass: Collect All Dates and History Events ---
-        for ln in lines:
-            # History events
-            hist_match = history_pattern.match(ln)
-            if hist_match:
-                history_str = hist_match.group(1)
-                event_blocks = re.findall(r'\{([^}]+)\}', history_str)
-                for block in event_blocks:
-                    details = {k.strip(): v.strip() for k, v in (pair.split('=', 1) for pair in block.split(',') if '=' in pair)}
-                    if 'When' in details:
-                        try:
-                            dt_obj = datetime.fromisoformat(details['When'].replace('Z', '+00:00'))
-                            data["history_events"].append((dt_obj, details))
-                        except (ValueError, IndexError):
-                            pass
-                continue
-
-            # Generic date lines
-            kv_match = _KV_PATTERN.match(ln)
-            if not kv_match: 
-                continue
-
-            val_str = kv_match.group("value").strip()
-            match = _DATE_TZ_PATTERN.match(val_str)
-            
-            if match:
-                parts = match.groupdict()
-                # Massage date into ISO format: YYYY-MM-DDTHH:MM:SS
-                date_part = parts.get("date").replace(":", "-", 2).replace(" ", "T")
-                tz_part = parts.get("tz")
-                
-                try:
-                    full_date_str = date_part
-                    if tz_part:
-                        full_date_str += tz_part.replace('Z', '+00:00')
-                    
-                    dt = datetime.fromisoformat(full_date_str)
-                    
-                    tag = kv_match.group("tag").strip().lower().replace(" ", "")
-                    group = kv_match.group("group").strip()
-                    data["all_dates"].append({"dt": dt, "tag": tag, "group": group, "full_str": val_str})
-
-                except ValueError:
-                    continue
-        
-        # Find first create date and latest modify date
-        for d in data["all_dates"]:
-            if d["tag"] in {"createdate", "creationdate"}:
-                if data["create_dt"] is None or d["dt"] < data["create_dt"]:
-                    data["create_dt"] = d["dt"]
-            elif d["tag"] in {"modifydate", "metadatadate"}:
-                if data["modify_dt"] is None or d["dt"] > data["modify_dt"]:
-                    data["modify_dt"] = d["dt"]
-        
-        return data
-        
-    def _detect_tool_change_from_exif(self, exiftool_output: str):
-        """
-        Determines if the primary tool changed between creation and last modification.
-        This function is now a lightweight wrapper around _parse_exif_data.
-        """
-        data = self._parse_exif_data(exiftool_output)
-        
-        create_tool = data["producer_pdf"] or data["producer_xmppdf"] or data["application"] or data["software"] or data["creatortool"] or ""
-        modify_tool = data["softwareagent"] or data["producer_pdf"] or data["producer_xmppdf"] or data["application"] or data["software"] or data["creatortool"] or ""
-        
-        create_engine = modify_engine = ""
-        if data["xmptoolkit"]:
-            if data["create_dt"]: create_engine = data["xmptoolkit"]
-            if data["modify_dt"]: modify_engine = data["xmptoolkit"]
-
-        changed_tool = bool(create_tool and modify_tool and create_tool.strip() != modify_tool.strip())
-        changed_engine = bool(create_engine and modify_engine and create_engine.strip() != modify_engine.strip())
-        
-        reason = ""
-        if changed_tool and changed_engine: reason = "mixed"
-        elif changed_tool: reason = "producer" if (data["producer_pdf"] or data["producer_xmppdf"]) else "software"
-        elif changed_engine: reason = "engine"
-
-        return {
-            "changed": changed_tool or changed_engine,
-            "create_tool": create_tool, "modify_tool": modify_tool,
-            "create_engine": create_engine, "modify_engine": modify_engine,
-            "modify_dt": data["modify_dt"],
-            "reason": reason
-        }
-
-    def _parse_exiftool_timeline(self, exiftool_output):
-        """
-        Generates a list of timeline events from parsed EXIF data.
-        """
-        events = []
-        data = self._parse_exif_data(exiftool_output)
-
-        create_tool = data["producer_pdf"] or data["producer_xmppdf"] or data["application"] or data["software"] or data["creatortool"] or ""
-        modify_tool = data["softwareagent"] or create_tool # Fallback to create_tool if no specific modify tool
-
-        # --- Add History Events ---
-        for dt_obj, details in data["history_events"]:
-            action = details.get('Action', 'N/A')
-            agent = details.get('SoftwareAgent', '')
-            changed = details.get('Changed', '')
-            desc = [f"Action: {action}"]
-            if agent: desc.append(f"Agent: {agent}")
-            if changed: desc.append(f"Changed: {changed}")
-            events.append((dt_obj, f"XMP History   - {' | '.join(desc)}"))
-
-        # --- Add Generic Date Events ---
-        def _ts_label(tag: str) -> str:
-            t = tag.replace(" ", "").lower()
-            return {"createdate": "Created", "creationdate": "Created", "modifydate": "Modified", "metadatadate": "Metadata"}.get(t, tag)
-
-        for d in data["all_dates"]:
-            label = self._(_ts_label(d["tag"]).lower())
-            tool = create_tool if d["tag"] in {"createdate", "creationdate"} else modify_tool
-            tool_part = f" | Tool: {tool}" if tool else ""
-            events.append((d["dt"], f"ExifTool ({d['group']}) - {label}: {d['full_str']}{tool_part}"))
-        
-        # --- Add XMP Engine Information ---
-        if data["xmptoolkit"]:
-            anchor_dt = data["create_dt"] or (data["all_dates"][0]["dt"] if data["all_dates"] else datetime.now())
-            label_engine = "XMP Engine" if self.language.get() == "en" else "XMP-motor"
-            events.append((anchor_dt, f"{label_engine}: {data['xmptoolkit']}"))
-
-        return events
         
     @staticmethod
     def _format_timedelta(delta):
@@ -3835,78 +3084,6 @@ class PDFReconApp:
 
         return f"(+{ ' '.join(parts) })"
             
-    def _parse_raw_content_timeline(self, file_content_string):
-        """Helper function to parse timestamps directly from the file's raw content."""
-        events = []
-        
-        # Extended PDF date pattern that captures optional timezone: D:YYYYMMDDHHmmss+HH'mm' or Z
-        pdf_date_extended = re.compile(
-            r"\/([A-Z][a-zA-Z0-9_]+)\s*\(\s*D:(\d{14})([+\-]\d{2}'\d{2}'|[+\-]\d{2}:\d{2}|[+\-]\d{4}|Z)?"
-        )
-        
-        for match in pdf_date_extended.finditer(file_content_string):
-            label, date_str, tz_str = match.groups()
-            try:
-                # Parse the base datetime
-                dt_obj = datetime.strptime(date_str, "%Y%m%d%H%M%S")
-                
-                # Handle timezone if present
-                if tz_str:
-                    if tz_str == 'Z':
-                        dt_obj = dt_obj.replace(tzinfo=timezone.utc)
-                    else:
-                        # Normalize: +01'00' or +01:00 or +0100 -> +01:00
-                        tz_clean = tz_str.replace("'", "").replace(":", "")
-                        if len(tz_clean) == 5:  # e.g., +0100
-                            tz_clean = tz_clean[:3] + ":" + tz_clean[3:]
-                        try:
-                            dt_obj = datetime.fromisoformat(dt_obj.strftime("%Y-%m-%dT%H:%M:%S") + tz_clean)
-                        except ValueError:
-                            pass  # Keep as naive if parsing fails
-                
-                tz_display = tz_str if tz_str else ""
-                display_line = f"Raw File: /{label}: {dt_obj.strftime('%Y-%m-%d %H:%M:%S')}{tz_display}"
-                events.append((dt_obj, display_line))
-            except ValueError:
-                continue
-
-        # Look for XMP-style dates: <xmp:CreateDate>2023-01-01T12:00:00Z</xmp:CreateDate>
-        xmp_date_pattern = re.compile(r"<([a-zA-Z0-9:]+)[^>]*?>\s*(\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}[^\s<]*)\s*<\/([a-zA-Z0-9:]+)>")
-        for match in xmp_date_pattern.finditer(file_content_string):
-            label, date_str, closing_label = match.groups()
-            if label != closing_label: 
-                continue
-            try:
-                # Normalize the date string for fromisoformat
-                # Handle Z -> +00:00, strip milliseconds but keep timezone
-                normalized = date_str.strip()
-                
-                # Check if there's timezone info
-                has_tz = 'Z' in normalized or '+' in normalized[10:] or (normalized.count('-') > 2 and '-' in normalized[10:])
-                
-                # Replace Z with +00:00 for fromisoformat
-                normalized = normalized.replace('Z', '+00:00')
-                
-                # Remove milliseconds but keep timezone
-                if '.' in normalized:
-                    dot_pos = normalized.index('.')
-                    # Find where timezone starts (+ or - after the dot)
-                    tz_start = -1
-                    for i, c in enumerate(normalized[dot_pos:]):
-                        if c in '+-':
-                            tz_start = dot_pos + i
-                            break
-                    if tz_start > 0:
-                        normalized = normalized[:dot_pos] + normalized[tz_start:]
-                    else:
-                        normalized = normalized[:dot_pos]
-                
-                dt_obj = datetime.fromisoformat(normalized)
-                display_line = f"Raw File: <{label}>: {date_str}"
-                events.append((dt_obj, display_line))
-            except (ValueError, IndexError):
-                continue
-        return events        
     def _clean_cell_value(self, value):
         """Fjerner tegn, der kan give XLSX/XML-fejl (BOM/mojibake/kontroltegn)."""
         import re
@@ -3925,250 +3102,6 @@ class PDFReconApp:
         s = s.replace("\x00", "")
         return s
 
-    def generate_comprehensive_timeline(self, filepath, raw_file_content, exiftool_output):
-        """
-        Combines events from all sources, separating them into timezone-aware and naive lists.
-        """
-        all_events = []
-
-        # 1) Get File System, ExifTool, and Raw Content timestamps
-        all_events.extend(self._get_filesystem_times(filepath))
-        all_events.extend(self._parse_exiftool_timeline(exiftool_output))
-        all_events.extend(self._parse_raw_content_timeline(raw_file_content))
-
-        # 2) Add a special event if a tool change was detected
-        try:
-            info = self._detect_tool_change_from_exif(exiftool_output)
-            if info.get("changed"):
-                when = info.get("modify_dt")
-                if not when and all_events:
-                    # Find a datetime object to anchor the event, prioritizing naive ones if present
-                    naive_dts = [e[0] for e in all_events if e[0].tzinfo is None]
-                    when = max(naive_dts) if naive_dts else max(e[0] for e in all_events)
-                if not when:
-                    when = datetime.now()
-                
-                # Format the description of the tool change
-                if self.language.get() == "da":
-                    label = "Værktøj skiftet"
-                    parts = [f"{info.get('create_tool','?')} → {info.get('modify_tool','?')}"]
-                    if info.get("reason") == "engine":
-                        parts.append(f"(XMP-motor: {info.get('create_engine','?')} → {info.get('modify_engine','?')})")
-                    line = f"{label}: " + " ".join(parts)
-                else:
-                    label = "Tool changed"
-                    parts = [f"{info.get('create_tool','?')} → {info.get('modify_tool','?')}"]
-                    if info.get("reason") == "engine":
-                        parts.append(f"(XMP engine: {info.get('create_engine','?')} → {info.get('modify_engine','?')})")
-                    line = f"{label}: " + " ".join(parts)
-                all_events.append((when, line))
-        except Exception:
-            pass
-
-        # 3) Separate events into two lists based on timezone info
-        aware_events = []
-        naive_events = []
-        for dt_obj, description in all_events:
-            if dt_obj.tzinfo is not None:
-                aware_events.append((dt_obj, description))
-            else:
-                naive_events.append((dt_obj, description))
-
-        # 4) Sort each list independently
-        aware_events.sort(key=lambda x: x[0])
-        naive_events.sort(key=lambda x: x[0])
-
-        return {"aware": aware_events, "naive": naive_events}
-    
-    @staticmethod
-    def decompress_stream(b):
-        """Attempts to decompress a PDF stream using common filters."""
-        for fn in (zlib.decompress, lambda d: base64.a85decode(re.sub(rb"\s", b"", d), adobe=True), lambda d: binascii.unhexlify(re.sub(rb"\s|>", b"", d))):
-            try: return fn(b).decode("latin1", "ignore")
-            except Exception: pass
-        return ""
-
-    @staticmethod
-    def extract_text(raw: bytes):
-        """
-        Extracts only what's needed for indicator hunting:
-        - ~2 MB header/trailer
-        - Small streams (skipping large image streams)
-        - XMP xpacket (if present)
-        """
-        txt_segments = []
-
-        # Cap: header/trailer/objects
-        head_cap = raw[:2_000_000].decode("latin1", "ignore")
-        txt_segments.append(head_cap)
-
-        # Only process small streams (e.g., <= 256 KB) to avoid inflating large images
-        for m in re.finditer(rb"stream\r?\n(.*?)\r?\nendstream", raw, re.S):
-            body = m.group(1)
-            if len(body) <= 256_000:
-                try:
-                    txt_segments.append(PDFReconApp.decompress_stream(body))
-                except Exception:
-                    try:
-                        txt_segments.append(body.decode("latin1", "ignore"))
-                    except Exception:
-                        pass
-
-        # XMP xpacket (full content)
-        m = re.search(rb"<\?xpacket begin=.*?\?>(.*?)<\?xpacket end=[^>]*\?>", raw, re.S)
-        if m:
-            try:
-                txt_segments.append(m.group(1).decode("utf-8", "ignore"))
-            except Exception:
-                txt_segments.append(m.group(1).decode("latin1", "ignore"))
-
-        # Ensure TouchUp_TextEdit is detectable even if it appears outside sampled text.
-        if re.search(rb"touchup_textedit", raw, re.I):
-            txt_segments.append("TouchUp_TextEdit")
-
-        return "\n".join(txt_segments)
-
-    def analyze_fonts(self, filepath, doc):
-            """
-            Analyzes fonts to detect multiple subsets of the same base font.
-            Returns a dictionary of conflicting fonts, e.g.,
-            {'Calibri': {'ABC+Calibri', 'DEF+Calibri-Bold'}}
-            """
-            font_subsets = {}
-            # Iterate through each page to get the fonts used
-            for page_num in range(len(doc)):
-                fonts_on_page = doc.get_page_fonts(page_num)
-                for font_info in fonts_on_page:
-                    basefont_name = font_info[3]
-                    if "+" in basefont_name:
-                        try:
-                            _, actual_base_font = basefont_name.split("+", 1)
-                            normalized_base = actual_base_font.split('-')[0]
-                            if normalized_base not in font_subsets:
-                                font_subsets[normalized_base] = set()
-                            font_subsets[normalized_base].add(basefont_name)
-                        except ValueError:
-                            continue
-            
-            # Filter for only those fonts that actually have multiple subsets
-            conflicting_fonts = {base: subsets for base, subsets in font_subsets.items() if len(subsets) > 1}
-            if conflicting_fonts:
-                logging.info(f"Multiple font subsets found in {filepath.name}: {conflicting_fonts}")
-
-            return conflicting_fonts
-    def detect_indicators(self, filepath, txt: str, doc):
-        """
-        Searches for indicators, now with an integrated text-diff feature for TouchUp edits.
-        """
-        indicators = {}
-
-        # --- High-Confidence Indicators ---
-        if re.search(r"touchup_textedit", txt, re.I):
-            found_text = self._extract_touchup_text(doc)
-            details = {'found_text': found_text, 'text_diff': None}
-
-            # Check for revisions to perform a text diff
-            revisions = self.extract_revisions(doc.write(), filepath)
-            if revisions:
-                latest_rev_path, _, latest_rev_bytes = revisions[0] # Compare against the most recent revision
-                logging.info(f"TouchUp found with revisions. Performing text diff for {filepath.name}.")
-                
-                original_text = self._get_text_for_comparison(filepath)
-                revision_text = self._get_text_for_comparison(latest_rev_bytes)
-
-                if original_text and revision_text:
-                    diff = difflib.unified_diff(
-                        revision_text.splitlines(keepends=True),
-                        original_text.splitlines(keepends=True),
-                        fromfile='Previous Version',
-                        tofile='Final Version',
-                    )
-                    details['text_diff'] = list(diff)
-            indicators['TouchUp_TextEdit'] = details
-
-        # --- Metadata Indicators ---
-        creators = set(re.findall(r"/Creator\s*\((.*?)\)", txt, re.I))
-        if len(creators) > 1:
-            indicators['MultipleCreators'] = {'count': len(creators), 'values': list(creators)}
-        
-        producers = set(re.findall(r"/Producer\s*\((.*?)\)", txt, re.I))
-        if len(producers) > 1:
-            indicators['MultipleProducers'] = {'count': len(producers), 'values': list(producers)}
-
-        if re.search(r'<xmpMM:History>', txt, re.I | re.S):
-            indicators['XMPHistory'] = {}
-
-        # --- Structural and Content Indicators ---
-        try:
-            conflicting_fonts = self.analyze_fonts(filepath, doc)
-            if conflicting_fonts:
-                indicators['MultipleFontSubsets'] = {'fonts': conflicting_fonts}
-        except Exception as e:
-            logging.error(f"Error analyzing fonts for {filepath.name}: {e}")
-
-        if (hasattr(doc, 'is_xfa') and doc.is_xfa) or "/XFA" in txt:
-            indicators['HasXFAForm'] = {}
-
-        if re.search(r"/Type\s*/Sig\b", txt):
-            indicators['HasDigitalSignature'] = {}
-
-        # --- Incremental Update Indicators ---
-        startxref_count = txt.lower().count("startxref")
-        if startxref_count > 1:
-            indicators['MultipleStartxref'] = {'count': startxref_count}
-        
-        prevs = re.findall(r"/Prev\s+\d+", txt)
-        if prevs:
-            indicators['IncrementalUpdates'] = {'count': len(prevs) + 1}
-        
-        if re.search(r"/Linearized\s+\d+", txt):
-            indicators['Linearized'] = {}
-            if startxref_count > 1 or prevs:
-                indicators['LinearizedUpdated'] = {}
-
-        # --- Feature Indicators ---
-        if re.search(r"/Redact\b", txt, re.I): indicators['HasRedactions'] = {}
-        if re.search(r"/Annots\b", txt, re.I): indicators['HasAnnotations'] = {}
-        if re.search(r"/PieceInfo\b", txt, re.I): indicators['HasPieceInfo'] = {}
-        if re.search(r"/AcroForm\b", txt, re.I):
-            indicators['HasAcroForm'] = {}
-            if re.search(r"/NeedAppearances\s+true\b", txt, re.I):
-                indicators['AcroFormNeedAppearances'] = {}
-
-        gen_gt_zero_matches = [m for m in re.finditer(r"\b(\d+)\s+(\d+)\s+obj\b", txt) if int(m.group(2)) > 0]
-        if gen_gt_zero_matches:
-            indicators['ObjGenGtZero'] = {'count': len(gen_gt_zero_matches)}
-
-        # --- ID Comparison ---
-        def _norm_uuid(x):
-            if x is None: return None
-            s = str(x).strip().upper()
-            return re.sub(r"^(URN:UUID:|UUID:|XMP\.IID:|XMP\.DID:)", "", s).strip("<>")
-
-        xmp_orig = _norm_uuid(re.search(r"xmpMM:OriginalDocumentID(?:>|=\")([^<\"]+)", txt, re.I).group(1) if re.search(r"xmpMM:OriginalDocumentID", txt, re.I) else None)
-        xmp_doc = _norm_uuid(re.search(r"xmpMM:DocumentID(?:>|=\")([^<\"]+)", txt, re.I).group(1) if re.search(r"xmpMM:DocumentID", txt, re.I) else None)
-        
-        if xmp_orig and xmp_doc and xmp_doc != xmp_orig:
-            indicators['XMPIDChange'] = {'from': xmp_orig, 'to': xmp_doc}
-
-        trailer_match = re.search(r"/ID\s*\[\s*<\s*([0-9A-Fa-f]+)\s*>\s*<\s*([0-9A-Fa-f]+)\s*>\s*\]", txt)
-        if trailer_match:
-            trailer_orig, trailer_curr = _norm_uuid(trailer_match.group(1)), _norm_uuid(trailer_match.group(2))
-            if trailer_orig and trailer_curr and trailer_curr != trailer_orig:
-                indicators['TrailerIDChange'] = {'from': trailer_orig, 'to': trailer_curr}
-        
-        # --- Date Mismatch ---
-        info_dates = dict(re.findall(r"/(ModDate|CreationDate)\s*\(\s*D:(\d{8,14})", txt))
-        xmp_dates = {k: v for k, v in re.findall(r"<xmp:(ModifyDate|CreateDate)>([^<]+)</xmp:\1>", txt)}
-
-        def _short(d: str) -> str: return re.sub(r"[-:TZ]", "", d)[:14]
-
-        if "CreationDate" in info_dates and "CreateDate" in xmp_dates and _short(info_dates["CreationDate"]) != _short(xmp_dates["CreateDate"]):
-            indicators['CreateDateMismatch'] = {'info': info_dates["CreationDate"], 'xmp': xmp_dates["CreateDate"]}
-        if "ModDate" in info_dates and "ModifyDate" in xmp_dates and _short(info_dates["ModDate"]) != _short(xmp_dates["ModifyDate"]):
-            indicators['ModifyDateMismatch'] = {'info': info_dates["ModDate"], 'xmp': xmp_dates["ModifyDate"]}
-        
-        return indicators
 
     def get_flag(self, indicators_dict, is_revision, parent_id=None):
         """
@@ -4425,193 +3358,8 @@ class PDFReconApp:
         close_button.grid(row=1, column=0, pady=(10, 0))
         
     
-    def _process_and_validate_revisions(self, potential_revisions, original_fp, scan_root_folder):
-        """
-        Processes, validates, and compares a list of potential PDF revisions.
-        Returns a list of dictionaries for valid revisions that should be reported.
-        """
-        valid_revision_results = []
-        invalid_xref_dir = scan_root_folder / "Invalid XREF"
-
-        for rev_path, basefile, rev_raw in potential_revisions:
-            tmp_path = None
-            try:
-                # Use a temporary file to run ExifTool without writing to the final destination first
-                with tempfile.NamedTemporaryFile(delete=False, suffix=".pdf") as tmp_file:
-                    tmp_file.write(rev_raw)
-                    tmp_path = Path(tmp_file.name)
-                
-                rev_exif = self.exiftool_output(tmp_path, detailed=True)
-
-                # Conditionally copy revisions with invalid XREF tables if the setting is enabled
-                if PDFReconConfig.EXPORT_INVALID_XREF and "Warning" in rev_exif and "Invalid xref table" in rev_exif:
-                    logging.info(f"Submitting invalid XREF revision for {basefile} to be copied.")
-                    dest_path = invalid_xref_dir / rev_path.name
-                    if self.copy_executor:
-                        self.copy_executor.submit(self._perform_copy, rev_raw, dest_path)
-                    continue  # Skip adding this invalid revision to the results
-
-                # If valid, write the revision to its final destination
-                rev_path.parent.mkdir(exist_ok=True)
-                rev_path.write_bytes(rev_raw)
-                
-                rev_md5 = hashlib.md5(rev_raw).hexdigest()
-                rev_txt = self.extract_text(rev_raw)
-                revision_timeline = self.generate_comprehensive_timeline(rev_path, rev_txt, rev_exif)
-
-                # Perform a visual comparison to see if the revision is identical to the original
-                is_identical = False
-                try:
-                    with fitz.open(original_fp) as doc_orig, fitz.open(rev_path) as doc_rev:
-                        pages_to_compare = min(doc_orig.page_count, doc_rev.page_count, PDFReconConfig.VISUAL_DIFF_PAGE_LIMIT)
-                        if pages_to_compare > 0:
-                            is_identical = True
-                            for i in range(pages_to_compare):
-                                page_orig, page_rev = doc_orig.load_page(i), doc_rev.load_page(i)
-                                if page_orig.rect != page_rev.rect: 
-                                    is_identical = False
-                                    break
-                                pix_orig, pix_rev = page_orig.get_pixmap(dpi=96), page_rev.get_pixmap(dpi=96)
-                                img_orig = Image.frombytes("RGB", [pix_orig.width, pix_orig.height], pix_orig.samples)
-                                img_rev = Image.frombytes("RGB", [pix_rev.width, pix_rev.height], pix_rev.samples)
-                                # Ensure images are same size before comparison
-                                if img_orig.size != img_rev.size:
-                                    is_identical = False
-                                    break
-                                if ImageChops.difference(img_orig, img_rev).getbbox() is not None: 
-                                    is_identical = False
-                                    break
-                except Exception as e:
-                    logging.warning(f"Could not visually compare {rev_path.name}, keeping it. Error: {e}")
-                
-                revision_row_data = { 
-                    "path": rev_path, "indicator_keys": {"Revision": {}}, "md5": rev_md5, "exif": rev_exif, 
-                    "is_revision": True, "timeline": revision_timeline, "original_path": original_fp, 
-                    "is_identical": is_identical, "status": "success"
-                }
-                valid_revision_results.append(revision_row_data)
-            finally:
-                if tmp_path and tmp_path.exists():
-                    os.remove(tmp_path)
-        
-        return valid_revision_results
-
-    def _process_single_file(self, fp, scan_root_folder):
-        """
-        Processes a single PDF file, submitting copy jobs to a dedicated thread pool.
-        """
-        try:
-            self.validate_pdf_file(fp)
-
-            raw = fp.read_bytes()
-            logging.info(f"Processing file: {fp.name} ({len(raw) / (1024*1024):.1f}MB)")
-            
-            # Use safer PDF opening
-            try:
-                doc = self._safe_pdf_open(fp, raw_bytes=raw)
-            except Exception as e:
-                logging.error(f"Failed to open PDF {fp.name}: {e}")
-                raise PDFCorruptionError(f"Cannot open PDF: {str(e)}")
-
-            # Extract text for indicator detection (must capture raw PDF structure)
-            logging.info(f"Extracting text from {fp.name}...")
-            txt = self.extract_text(raw)
-            logging.info(f"Text extraction complete for {fp.name}: {len(txt)} characters")            # Detect indicators with error handling
-            logging.info(f"Detecting indicators in {fp.name}...")
-            try:
-                indicators = self.detect_indicators(fp, txt, doc)
-                logging.info(f"Indicator detection complete for {fp.name}")
-            except Exception as e:
-                logging.warning(f"Error detecting indicators in {fp.name}: {e}")
-                indicators = {}
-            
-            logging.info(f"Computing MD5 and EXIF for {fp.name}...")
-            md5_hash = md5_file(fp)
-            exif = self.exiftool_output(fp, detailed=True)
-
-            # Conditionally submit invalid XREF originals to the copy pool if setting is enabled
-            if PDFReconConfig.EXPORT_INVALID_XREF and "Warning" in exif and "Invalid xref table" in exif:
-                invalid_xref_dir = scan_root_folder / "Invalid XREF"
-                dest_path = invalid_xref_dir / fp.name
-                if self.copy_executor:
-                    self.copy_executor.submit(self._perform_copy, fp, dest_path)
-
-            tool_change_info = self._detect_tool_change_from_exif(exif)
-            if tool_change_info.get("changed"):
-                indicators['ToolChange'] = {}
-            
-            logging.info(f"Generating timeline for {fp.name}...")
-            original_timeline = self.generate_comprehensive_timeline(fp, txt, exif)
-            
-            logging.info(f"Extracting revisions from {fp.name}...")
-            potential_revisions = self.extract_revisions(raw, fp)
-            doc.close()
-            
-            self._add_layer_indicators(raw, fp, indicators)
-
-            # Delegate all revision processing to the new helper method
-            valid_revision_results = self._process_and_validate_revisions(potential_revisions, fp, scan_root_folder)
-
-            if valid_revision_results:
-                indicators['HasRevisions'] = {'count': len(valid_revision_results)}
-
-            original_row_data = {
-                "path": fp, "indicator_keys": indicators, "md5": md5_hash, 
-                "exif": exif, "is_revision": False, "timeline": original_timeline, "status": "success"
-            }
-            
-            results = [original_row_data] + valid_revision_results
-            return results
-
-        except PDFTooLargeError as e:
-            return [self._handle_file_processing_error(fp, "file_too_large", e)]
-        except PDFEncryptedError as e:
-            return [self._handle_file_processing_error(fp, "file_encrypted", e)]
-        except PDFCorruptionError as e:
-            return [self._handle_file_processing_error(fp, "file_corrupt", e)]
-        except Exception as e:
-            logging.exception(f"Unexpected error processing file {fp.name}")
-            return [self._handle_file_processing_error(fp, "processing_error", e)]
 
 
-            # Conditionally submit invalid XREF originals to the copy pool if setting is enabled
-            if PDFReconConfig.EXPORT_INVALID_XREF and "Warning" in exif and "Invalid xref table" in exif:
-                invalid_xref_dir = scan_root_folder / "Invalid XREF"
-                dest_path = invalid_xref_dir / fp.name
-                if self.copy_executor:
-                    self.copy_executor.submit(self._perform_copy, fp, dest_path)
-
-            tool_change_info = self._detect_tool_change_from_exif(exif)
-            if tool_change_info.get("changed"):
-                indicators['ToolChange'] = {}
-            original_timeline = self.generate_comprehensive_timeline(fp, txt, exif)
-            potential_revisions = self.extract_revisions(raw, fp)
-            doc.close()
-            
-            self._add_layer_indicators(raw, fp, indicators)
-
-            # Delegate all revision processing to the new helper method
-            valid_revision_results = self._process_and_validate_revisions(potential_revisions, fp, scan_root_folder)
-
-            if valid_revision_results:
-                indicators['HasRevisions'] = {'count': len(valid_revision_results)}
-            original_row_data = {
-                "path": fp, "indicator_keys": indicators, "md5": md5_hash, 
-                "exif": exif, "is_revision": False, "timeline": original_timeline, "status": "success"
-            }
-            
-            results = [original_row_data] + valid_revision_results
-            return results
-
-        except PDFTooLargeError as e:
-            return [self._handle_file_processing_error(fp, "file_too_large", e)]
-        except PDFEncryptedError as e:
-            return [self._handle_file_processing_error(fp, "file_encrypted", e)]
-        except PDFCorruptionError as e:
-            return [self._handle_file_processing_error(fp, "file_corrupt", e)]
-        except Exception as e:
-            logging.exception(f"Unexpected error processing file {fp.name}")
-            return [self._handle_file_processing_error(fp, "processing_error", e)]
     def _prompt_and_export(self, file_format):
         """Prompts the user for a file path and calls the relevant export function."""
         if not self.report_data:
@@ -4847,154 +3595,6 @@ class PDFReconApp:
                 rows=rows
             ))
    
-    def _extract_touchup_text(self, doc):
-        """
-        Parses a PDF's internal objects to find any text associated with a TouchUp_TextEdit flag.
-        Returns a list of extracted text strings.
-        """
-        def _clean_text_segment(raw_bytes):
-            try:
-                text = raw_bytes.decode("latin-1", errors="ignore")
-                replacements = {"Õ": "å", "ã": "å", "°": "ø", "¯": "ø", "µ": "æ", "\xa0": " "}
-                for old, new in replacements.items():
-                    text = text.replace(old, new)
-                text = re.sub(r"[\x00-\x1f\x7f-\x9f]", "", text)
-                text = re.sub(r" +", " ", text)
-                return text.strip()
-            except Exception:
-                return ""
-
-        def _is_probably_junk(text):
-            if not text or len(text) < 3:
-                return True
-            allowed = 0
-            for ch in text:
-                if ch.isalnum() or ch.isspace() or ch in ".,:;!?-()'\"/":
-                    allowed += 1
-            if allowed / max(len(text), 1) < 0.7:
-                return True
-            if len(set(text)) <= 3 and len(text) > 10:
-                return True
-            return False
-
-        # First try pikepdf-based extraction (TouchUp blocks).
-        try:
-            import pikepdf
-            from pikepdf import String
-            from io import BytesIO
-
-            results = []
-            pdf = None
-            try:
-                original_filepath = doc.name if doc and hasattr(doc, "name") else None
-                if original_filepath and Path(original_filepath).exists():
-                    pdf = pikepdf.open(original_filepath)
-                elif doc and not doc.is_closed and hasattr(doc, "write"):
-                    pdf = pikepdf.open(BytesIO(doc.write()))
-            except Exception as e:
-                logging.debug(f"Pikepdf open failed for TouchUp extraction: {e}")
-                pdf = None
-
-            if pdf is not None:
-                # Group results by page number: {page_num: [text1, text2, ...]}
-                page_results = {}
-                with pdf:
-                    for i, page in enumerate(pdf.pages):
-                        page_num = i + 1
-                        try:
-                            commands = pikepdf.parse_content_stream(page)
-                        except Exception:
-                            continue
-
-                        active_search = False
-                        current_block_buffer = []
-
-                        for operands, operator in commands:
-                            op_name = str(operator)
-                            if any("TouchUp" in str(arg) for arg in operands):
-                                active_search = True
-                            elif active_search and op_name in ["Tj", "TJ"]:
-                                chunks = operands[0] if op_name == "TJ" else operands
-                                for chunk in chunks:
-                                    if isinstance(chunk, String):
-                                        try:
-                                            raw_text = chunk.decode()
-                                            cleaned = _clean_text_segment(raw_text.encode("latin-1", errors="ignore"))
-                                        except Exception:
-                                            cleaned = _clean_text_segment(bytes(chunk))
-                                        if cleaned and not _is_probably_junk(cleaned):
-                                            current_block_buffer.append(cleaned)
-                            elif op_name in ["ET", "EMC"]:
-                                if active_search and current_block_buffer:
-                                    # Join with visible separator so user can see segment boundaries
-                                    # Use │ (box drawing character) as delimiter
-                                    combined = " │ ".join([b for b in current_block_buffer if b])
-                                    if combined:
-                                        if page_num not in page_results:
-                                            page_results[page_num] = []
-                                        page_results[page_num].append(combined)
-                                    current_block_buffer = []
-                                active_search = False
-
-                if page_results:
-                    # Return as dictionary grouped by page
-                    return page_results
-        except ImportError:
-            logging.debug("pikepdf not installed, falling back to legacy TouchUp extraction.")
-        except Exception as e:
-            logging.warning(f"TouchUp pikepdf extraction failed: {e}")
-
-        # Fallback: legacy extraction from xref streams.
-        # Returns dict format: {0: [text1, text2, ...]} (page 0 = unknown page)
-        found_text = {}
-        if not doc or doc.is_closed:
-            return found_text
-            
-        for xref in range(1, doc.xref_length()):
-            try:
-                obj_source = doc.xref_object(xref, compressed=False)
-                if "/TouchUp_TextEdit" in obj_source:
-                    stream = doc.xref_stream(xref)
-                    if stream:
-                        matches = re.findall(rb"\((.*?)\)\s*Tj", stream)
-                        for match in matches:
-                            try:
-                                decoded_text = fitz.utils.pdfdoc_decode(match)
-                                if decoded_text.strip():
-                                    # Use page 0 as "unknown page" for legacy extraction
-                                    if 0 not in found_text:
-                                        found_text[0] = []
-                                    found_text[0].append(decoded_text.strip())
-                            except Exception:
-                                continue
-            except Exception as e:
-                logging.warning(f"Could not process object {xref} for TouchUp text: {e}")
-                continue
-        return found_text
-    
-    def _get_text_for_comparison(self, source):
-        """
-        Performs a robust, layout-preserving text extraction on a PDF.
-        Source can be bytes, a string path, or a Path object.
-        """
-        full_text = []
-        doc = None
-        try:
-            if isinstance(source, bytes):
-                doc = fitz.open(stream=source, filetype="pdf")
-            else:
-                resolved_path = self._resolve_case_path(source) # Resolve the path
-                doc = fitz.open(resolved_path)
-
-            for page in doc:
-                full_text.append(page.get_text("text", sort=True))
-            return "\n".join(full_text)
-        except Exception as e:
-            logging.error(f"Robust text extraction failed: {e}")
-            return ""
-        finally:
-            if doc:
-                doc.close()    
         
             
     def _get_touchup_regions_for_page(self, fitz_doc, page_num, touchup_texts):
